% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{datamuse_api}
\alias{datamuse_api}
\title{Call Datamuse API and return data.}
\usage{
datamuse_api(path, limit = 10)
}
\arguments{
\item{path}{path to append to Datamuse API endpoint.}

\item{limit}{number of results to limit the API response to.}
}
\value{
data returned from API call.
}
\description{
Call Datamuse API and return data.
}
\examples{
datamuse_api("/words?rel_rhy=test")

datamuse_api("/words?ml=test")
}
