% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_content}
\alias{get_content}
\title{Extract content dataframe from datamuse API call.}
\usage{
get_content(full_path, limit = 10)
}
\arguments{
\item{full_path}{API path to append to datamuse API endpoint.}

\item{limit}{max number of rows to return from the content dataframe.}
}
\value{
dataframe with content returned from API call.
}
\description{
Extract content dataframe from datamuse API call.
}
\examples{
get_content("/words?rel_rhy=test", limit = 5)
get_content("/words?ml=test", limit = 20)
}
