% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_means_like}
\alias{get_means_like}
\title{Get dataframe of words with similar meaning.}
\usage{
get_means_like(word = "test", limit = 10)
}
\arguments{
\item{word}{to get similar meaning with.}

\item{limit}{max number of words to return.}
}
\value{
dataframe with words of similar meaning.
}
\description{
Get dataframe of words with similar meaning.
}
\examples{
get_means_like("test")
get_means_like("test", limit = 10)
}
