% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_other_related}
\alias{get_other_related}
\title{Get dataframe of words that are related according to a supplied code.}
\usage{
get_other_related(word = "test", code = "jja", limit = 10)
}
\arguments{
\item{word}{to get similarly spelled words with.}

\item{code}{related word code from http://www.datamuse.com/api/.}

\item{limit}{max number of words to return.}
}
\value{
dataframe with words that are related.
}
\description{
Get dataframe of words that are related according to a supplied code.
}
\examples{
get_other_related("test", code = "jja", limit = 3)
get_other_related("test", code = "cns", limit = 10)
}
