% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_rhyme}
\alias{get_rhyme}
\title{Get dataframe of rhyming words.}
\usage{
get_rhyme(word, limit = 10)
}
\arguments{
\item{word}{to rhyme with.}

\item{limit}{max number of words to return\.}
}
\value{
dataframe with rhyming words.
}
\description{
Get dataframe of rhyming words.
}
\examples{
get_rhyme("test")
get_rhyme("test", limit = 10)
}
