% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_sounds_like}
\alias{get_sounds_like}
\title{Get dataframe of words that sound similar.}
\usage{
get_sounds_like(word = "test", limit = 10)
}
\arguments{
\item{word}{to get similar sounding words with.}

\item{limit}{max number of words to return.}
}
\value{
dataframe with words that sound similar.
}
\description{
Get dataframe of words that sound similar.
}
\examples{
get_sounds_like("test")
get_sounds_like("test", limit = 10)
}
