% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_spelled_like}
\alias{get_spelled_like}
\title{Get dataframe of words that are spelled similarly.}
\usage{
get_spelled_like(word = "test", limit = 10)
}
\arguments{
\item{word}{to get similarly spelled words with.}

\item{limit}{max number of words to return.}
}
\value{
dataframe with words that are spelled similarly.
}
\description{
Get dataframe of words that are spelled similarly.
}
\examples{
get_spelled_like("test")
get_spelled_like("test", limit = 10)
}
