% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{print.datamuse_api}
\alias{print.datamuse_api}
\title{Print method for data returned by API.}
\usage{
\method{print}{datamuse_api}(x, ...)
}
\arguments{
\item{x}{datamuse API return object.}

\item{...}{the other stuff.}
}
\description{
Print method for data returned by API.
}
\examples{
print(datamuse_api("/words?rel_rhy=test"))
print(datamuse_api("/words?ml=test"))
}
