% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{return_content}
\alias{return_content}
\title{Helper function to shape data to return to user.}
\usage{
return_content(api_content, return_type)
}
\arguments{
\item{api_content}{content returned from the API call.}

\item{return_type}{type to return. Options are:
\itemize{
\item 'df' for data.frame.
\item 'word' for a single word.
\item 'random_word' or 'random word' or 'rand' for a random word.
\item 'vector' for a vector of words.
}}
}
\value{
data as specified above.
}
\description{
Helper function to shape data to return to user.
}
