% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_means_like}
\alias{get_means_like}
\title{Get dataframe for words with similar meaning.}
\usage{
get_means_like(word = "test", return_type = "df", limit = 10)
}
\arguments{
\item{word}{to get similar meaning with.}

\item{return_type}{type to return. Options are:
\itemize{
\item 'df' for data.frame.
\item 'word' for a single word.
\item 'random_word' or 'random word' or 'rand' for a random word.
\item 'vector' for a vector of words.
}}

\item{limit}{max number of words to return.}
}
\value{
data with words of similar meaning.
}
\description{
Get dataframe for words with similar meaning.
}
\examples{
get_means_like("test")

get_means_like("test", limit = 10)
}
