% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_sounds_like}
\alias{get_sounds_like}
\title{Get data for words that sound similar.}
\usage{
get_sounds_like(word = "test", return_type = "df", limit = 10, num_syl = NULL)
}
\arguments{
\item{word}{to get similar sounding words with.}

\item{return_type}{type to return. Options are:
\itemize{
\item 'df' for data.frame.
\item 'word' for a single word.
\item 'random_word' or 'random word' or 'rand' for a random word.
\item 'vector' for a vector of words.
}}

\item{limit}{max number of words to return.}

\item{num_syl}{number of syllables in rhymes to return.}
}
\value{
data containing word(s) that sound similar.
}
\description{
Get data for words that sound similar.
}
\examples{
get_sounds_like("test")

get_sounds_like("test", limit = 10)
}
