% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymer.R
\name{get_spelled_like}
\alias{get_spelled_like}
\title{Get data for words that are spelled similarly.}
\usage{
get_spelled_like(word = "test", return_type = "df", limit = 10)
}
\arguments{
\item{word}{to get similarly spelled words with.}

\item{return_type}{type to return. Options are:
\itemize{
\item 'df' for data.frame.
\item 'word' for a single word.
\item 'random_word' or 'random word' or 'rand' for a random word.
\item 'vector' for a vector of words.
}}

\item{limit}{max number of words to return.}
}
\value{
data with words that are spelled similarly.
}
\description{
Get data for words that are spelled similarly.
}
\examples{
get_spelled_like("test")

get_spelled_like("test", limit = 10)
}
